// src/interfaces/timeInformation.ts
export interface ITimeInformation {
    Id: number;
    Employee_ID_Number: number;
    Wroked_Hours: number;
    Off_Hours: number;
    Days_off: number;
    Over_Time: number;
    Extra_Days: number;
    w_From_Date_Day: number;
    w_From_Date_Month: number;
    w_From_Date_Year: number;
    w_To_Date_Day: number;
    w_To_Date_Month: number;
    w_To_Date_Year: number;
}